Created new file and added form to take in details

+++ models/transactions.rb
@@ -26,7 +26,7 @@ class Transaction
   def initialize(options)
     # Creates new instances of transactions, takes in hash of options
     @id = options['id'].to_i
-    @amount = options['amount'].to_i
+    @amount = options['amount'].to_f
     @merchant_id = options['merchant_id'].to_i
     @tag = options['tag']
   end
